requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
elasticsearch==8.13.0
python-dotenv==1.1.1
openai==1.54.0
google-auth==2.23.4
requests==2.31.0